HTTP Requests working, relations updated

+++ src/main/java/com/revature/repositories/CarRepo.java
@@ -11,7 +11,7 @@ import com.revature.model.Dealership;
 @Repository
 public interface CarRepo extends JpaRepository <Car, Long> {
 
-    public Car findById(long id);
+    public Car findById(int id);
     public List<Car> findCarsByMake(String make);
     public List<Car> findCarsByModel(String model);
     public List<Car> findCarsByYear(int year);
